refactor: Register types with scheme using pattern from Cluster API (#1124)
**What problem does this PR solve?**:
This pattern removes the dependency on controller-runtime from the API
module. It is recommended by Cluster API.


https://cluster-api.sigs.k8s.io/developer/providers/getting-started/implement-api-types#registering-apis-in-the-scheme

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
Because the api module contains external types that do have a dependency
on controller-runtime, we have to make a similar change to these types
upstream before we can remove our dependency on controller-runtime.