feat: add registry addon (#1116)
**What problem does this PR solve?**:
This PR adds a new addon `registryMirror` that deploys
https://github.com/distribution/distribution as a StatefulSet
and a [sidecar container to sync
images](https://github.com/regclient/regclient/blob/main/docs/regsync.md)
across instances.
 
```
$ kubectl get pods -n registry-system 
NAME                                           READY   STATUS    RESTARTS   AGE
cncf-distribution-registry-docker-registry-0   2/2     Running   0          2m13s
cncf-distribution-registry-docker-registry-1   2/2     Running   0          8s

```
 
This addon is designed to only be a mirror and not used a regular
registry, hence the name and the lack of external access to the Service.
In a follow up PR, the in-cluster Service will be used a Containerd
mirror.
In a follow up PR, it will also be deployed with randomly generated
credentials to further prevent direct use.
 
**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->