fixup! feat: Add preflight checks framework
Remove side effects from the initialization. That is, the checker
initialization still decides which checks apply, but we defer side
effects, and potential errors, to the checks themselves. This allows us
to execute all checks that apply, and get the results to the client.

Previously, if initialization failed, the checker returned no checks.