feat: generate a self-signed cert for registry addon (#1127)
**What problem does this PR solve?**:
This PR adds the functionality to set a self-signed certificate for the
registry addon.
A new `Certificate` deployed by the Helm chart generates a 10 year CA
Secret `registry-addon-root-ca` in the namespace where this controller
is running.

The `ca.crt` data is copied into a Cluster specific Secret (with an
OwnerRef set) in a `BeforeClusterCreate` hook and is then used by the
mirror handler.

This CA is then used to sign a new 2 year TLS certificate for every
cluster that is copied to the remote cluster
`registry-system/registry-tls` and used by the registry Pods. A new cert
is generated and copied on every invocation of the handler, on
`BeforeClusterUpgrade` and `AfterControlPlaneInitialized` lifecycle
hooks, so that we can punt on auto renewal for now.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
1. Tested manually by create a Docker cluster and verifying I can push
an image into the registry, observe it get synced across the replicas,
and then get pulled with Containerd.

```
$ kubectl get secret -n registry-system
NAME                                                TYPE                 DATA   AGE
cncf-distribution-registry-docker-registry-secret   Opaque               3      5h16m
registry-tls                                        kubernetes.io/tls    3      5h16m
sh.helm.release.v1.cncf-distribution-registry.v1    helm.sh/release.v1   1      5h16m


$ kubectl port-forward --address=127.0.0.1 --namespace registry-system pod/cncf-distribution-registry-docker-registry-0 5000:5000
$ crane copy nginx:stable 127.0.0.1:5000/library/nginx:dkoshkin --insecure
$ kubectl run nginx-working --image=docker.io/library/nginx:dkoshkin
$ kubectl get pods
NAME                                                              READY   STATUS              RESTARTS   AGE
cluster-autoscaler-0196db54-35b0-73fd-ad5b-14f998751820-7bzwnm9   0/1     ContainerCreating   0          5h15m
nginx-working                                                     1/1     Running             0          53m
```

2. New integration tests.
3. Existing e2e tests already wait for the STS Pods to be Ready which
won't happen unless the TLS Secret is there and is valid.

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
I've considered other approaches here:
1. Using cert-manager to generate a unique CA per cluster. This makes it
more difficult for clients to trust the registries when pushing images
to multiple clusters.
2. Use cert-manager to also generate the TLS certificate, it got pretty
complicated to trigger a renewal and needing to wait for it to reconcile
the cert Secret. Using cert-manager was also not providing much value as
we want better control when the certs get rotated -
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/commit/118b39e1b3b7a6dba6f070e1e3cb511a70398318.
3. Use the CAPI generated CA cert to sign a new cert, but decided
against using an externally managed CA for a separate usecase.