fix: update CNCF registry version to 2.3.4, app version 2.8.3 (#1150)
**What problem does this PR solve?**:
This version picks up CVE fixes in the registry image.
(Updating to 3.0 will be handled as a separate PR.)

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->