feat: Add preflight checks framework (#1129)
**What problem does this PR solve?**:
Adds a framework for preflight checks. A preflight check is a type of
validation that typically requires access to an infrastructure API.

A validating webhook on the Cluster resource executes all preflight
checks, and returns failures, and warnings to the client.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->