feat: Nutanix VM image preflight check (#1130)
**What problem does this PR solve?**:
Implements a preflight check that verifies the Nutanix VM images
referenced by the Cluster spec.

This also includes Nutanix Configuration and Credentials checks,
both of which are provide dependencies required by the VM images check.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
Stacked on #1129 

I am working on unit tests. The Nutanix client is difficult to mock, so
creating the tests is taking more time than I expected.