refactor: add global feature.Gates variable (#1167)
**What problem does this PR solve?**:
Refactoring the feature gate code to make it easier to use.
Added a new `feature.Gate` readonly variable that can be used anywhere
in the codebase to determine if a specific feature gate was set.

This is based on CAPI's implementation
https://github.com/kubernetes-sigs/cluster-api/tree/main/feature

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->