feat: Enforce MD replicas within cluster autoscaler bounds (#1169)
When the current MD `.spec.replicas` is outside of the bounds of the
cluster autoscaler min/max annotations, cluster autoscaler will not
scale the MD up or down. This can occur when either the annotations are
edited or `spec.replicas` is updated.

In a topology-based cluster, CAPI propagates the cluster autoscaler
annotations from
`cluster.spec.topology.workers.machineDeployments[].metadata.annotations`
to the templated `MachineDeployment` resources. If these annotations are
changed so that they do not contain the current number of replicas, then
no scaling up or down will happen to the MachineDeployment by the
cluster autoscaler.

This commit adds a controller that does the following to
MachineDeployments:

- if `spec.replicas` is nil do nothing
- if cluster autoscaler annotations for min or max are missing do
nothing
- if `spec.replicas` is within cluster autoscaler annotations bounds do
nothing
- if `spec.replicas` is not with CA annotations bounds then set
`spec.replicas`
to nil (CAPI MD defaulting webhook will correctly set MD replicas as per

https://github.com/kubernetes-sigs/cluster-api/blob/v1.10.3/internal/webhooks/machinedeployment.go\#L353-L376.

This ensures that scaling up and down by specifying cluster autoscaler
annotations on the machine deployments in
`cluster.spec.topology.workers.machineDeployments[].metadata.annotations`
will always be respected, even if the current number of replicas is
outside the autoscaling specified bounds.