fix: preserve registry addon root CA on move (#1155)
**What problem does this PR solve?**:
In the initial
[PR](https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1127)
I misunderstood how clusterctl move handles Secrets.

It only moves the Secrets in infra provider namespace and not the CAREN
namespace
https://cluster-api.sigs.k8s.io/developer/providers/contracts/clusterctl#move.
This results in the root CA being recreated and the management cluster
ends up using a different CA for the registry addon as compared to the
workload clusters.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
- Tested manually to verify that I can still push images.
- Added unit, integration and e2e tests.

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->