fix: Do not run preflight checks if Cluster is paused (#1181)
**What problem does this PR solve?**:
When clusterctl moves the Cluster, clusterctl first creates a paused
Cluster object, then the referenced Secrets. That order might be a bug.
However, to allow move to work, we want to skip preflight checks.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->