feat: auto enable registry addon in workload clusters (#1175)
**What problem does this PR solve?**:
This adds a new feature gate `AutoEnableWorkloadClusterRegistry=true`
that will automatically enable the registry addon on workload cluster
creation when it is enabled on the management cluster.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->