fix: misc fixes to the preflight framework (#1188)
**What problem does this PR solve?**:
- Skip preflight on Cluster update. This is defensive programming. The
webhook is already configured to only run on create. When we support for
update, we will remove this, and reconfigure the webhook.
- Use preflight check name as cause type. Previously, the
"FailedPreflight" prefix was present, but it makes it more difficult to
identify the name of the check that failed, and the prefix is not
necessary, because the top-level Message explains that preflight checks
failed.
- Debug log message when preflight check runs.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->