fix: add tolerations and nodeAffinity overrides for registry addon (#1183)
**What problem does this PR solve?**:
Adds antiAffinity to ensure that registry pods are scheduled on the
control plane nodes.
Adds tolerations and nodeAffinity to the registry pods to ensure that
they are scheduled on the control plane nodes.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
- E2E tests
- Manual tests

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->