fix: Do not treat expected preflight check failures as internal errors (#1195)
**What problem does this PR solve?**:
Internal errors are meant to signal some unexpected failure. We have
been treating a number of expected failures as internal errors, and this
PR fixes these.

It also makes CheckResult interface easier to understand by renaming
the Error field to InternalError.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->