fix: Fix typo in field name; use one-line strings to prevent future typos (#1206)
**What problem does this PR solve?**:
There was an extra period in a field, introduced when I manually split
the field across two lines, to appease the linter. Fields are long
strings, and easier to read and maintain if we don't split them up.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->