fix(auto-cert-renewal): adds 0 as valid value for daysBeforeExpiry (#1218)
**What problem does this PR solve?**:
As of now to disable CAPI based automated certificate renewal we have to
set `autoRenewCertificates` to `nil` which results in SSA patch
[issue](https://github.com/kubernetes/kubernetes/issues/117447). To
avoid this added 0 as a valid value for
`autoRenewCertificates.daysBeforeExpiry` denoting the feat is disabled.

**Which issue(s) this PR fixes**:
Fixes [NCN-108329](https://jira.nutanix.com/browse/NCN-108329)

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
- Tested using the dev workflow along with the konvoy
[changes](https://github.com/mesosphere/konvoy2/pull/3959)

---------

Co-authored-by: Dimitri Koshkin <dimitri.koshkin@nutanix.com>