feat: Add more context to MetalLB config apply conflict errors (#1231)
**What problem does this PR solve?**:
Backport of #1225

When there is a conflict writing the MetalLB configuration, we want to
give more information to the user.

**Which issue(s) this PR fixes**:
Fixes https://jira.nutanix.com/browse/NCN-107584

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes. Provide
output from the tests and any manual steps needed to replicate the
tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers. This
may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc. -->

---------

**What problem does this PR solve?**:

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->

Co-authored-by: Jimmi Dyson <jimmidyson@gmail.com>