fix: Correctly handle variable overrides
Currently the metamutator only passes the override values when they are
present on a machine deployment or control plane config. This commit
fixes that by merging the global variables into the overrides, resulting
in the behaviour that we always expected.

As overrides are not that commonly used yet, we have not seen this issue
but as taints, etc become more commonly used this will become an issue.

Tests added to prove the fix as well.