feat: EKS ClusterClass and Cluster example
The bases are included in the hack directory rather than pulling in
from external projects as we do for other providers because the EKS
CC support has not been released in CAPA yet. Once that has happened,
we can switch to the same strategy and use the upstream CC and Cluster
as the base for patching to create the CAREN CC/Cluster examples.