fix: hardcode the kubectl tag in registry syncer Pods (#1262)
**What problem does this PR solve?**:
Supersedes
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1261

Hardcode a version of the kubectl images so that its not dependent on
the Cluster version which can be a pre-release or have additional build
metadata.

We will periodically update this version.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->