perf: Enable compression of audit logs, and increase max file size
We enable compression of audit logs. Whenever a log file is rotated, it
is compressed.

Experiments show that gzip yields a compression factor of approximately
12 for audit logs, so we increase the maximum file size by that factor.

With the previous configuration, we had 1 log file of 100MB, and 10
rotated log files of 100MB each, for a total of 1100MB. That's our
target disk space requirement.

We want to keep this disk space usage, but enable compression. We can't
just make the maximum file size larger. For example, if we increase it
to 1000MB, then we'll end up with 1 1000MB log, plus 10 compressed,
rotated logs of approximately 100MB each, for a total of 2000MB, much
higher than our target.

To increase retention, we reduce the size of the uncompressed log. We
keep --audit-log-maxsize unchanged at 100, but increase
--audit-log-maxbackup to 90. We end up with 1 100MB uncompressed log,
plus 90 compressed, rotated logs of approximately 10MB each, for a total
of 1000MB. That's a little below our target!