feat: update AWS CSI to 2.48.0 (#1280)
**What problem does this PR solve?**:
Updates to latest version of AWS CSI.

Pulling the commit out of
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1266
to make its review easier.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->