ci: Use DeterminateSystems/nix-installer-action instead of cachix action (#1284)
**What problem does this PR solve?**:

This is the preferred upstream action and once it is incorporated into
devbox-install-action this can be removed from here too.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->