feat: Pull in EKS APIs (#1253)
**What problem does this PR solve?**:

Adds EKS APIs to external to start work on EKS support in CAREN.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
This depends on upstream changes that have been merged but not released
so we need to wait until we get a CAPA release that includes the EKS
ClusterClass support.