fixup! refactor: Use CAAPH templating rather than custom temlating
Use the ControlPlane reference in the CAAPH template and check the annotation
for skip kube proxy on the referenced control plane object. As per
https://cluster-api.sigs.k8s.io/reference/api/metadata-propagation\#cluster-topology
the skip proxy annotation will be propagated to the annotations on the
referenced control plane so is safe to check. CAAPH adds the referenced
control plane to the variables referencable in the Helm values template.