refactor: keep only common types in Generic config specs (#1297)
**What problem does this PR solve?**:
After EKS types were introduced, the fields in
`GenericClusterConfigSpec` and `GenericNodeSpec` are no longer all
"generic" because some do not apply to EKS clusters, as it does not use
kubeadm to create the cluster.

This PR attempts to address this by only keeping the fields that can be
set for both EKS and all other infra types in `GenericClusterConfigSpec`
and moves the rest of the fields to `KubeadmClusterConfigSpec`, and
similarly for `GenericNodeSpec`.

Because the structs are included `inline` in the individual Cluster
types this does not change the external yaml API, only how they are
referenced in Go (notice how the CRDs for
<aws|docker|nutanix>clusterconfigs did not change).

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->
Unit tests.

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->
If we agree with this approach I will have separate refactor PRs to
reorganize the handlers and the public docs.