fix: include Generic config specs in EKS cluster API  (#1298)
**What problem does this PR solve?**:
Stacked on
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1297

This fixes the generic handlers for EKS clusters by properly exposing
the `GenericClusterConfigSpec` API in `EKSClusterConfig`.

The following fields are still aspirational and are not properly
supported in the handlers for EKS, but they can be.
```
	// +kubebuilder:validation:Optional
	Proxy *HTTPProxy `json:"proxy,omitempty"`

	// +kubebuilder:validation:Optional
	// +kubebuilder:validation:MaxItems=32
	ImageRegistries []ImageRegistry `json:"imageRegistries,omitempty"`

	// +kubebuilder:validation:Optional
	GlobalImageRegistryMirror *GlobalImageRegistryMirror `json:"globalImageRegistryMirror,omitempty"`
```

But these are working and I've added broken tests to verify before
fixing the API.
```
	// +kubebuilder:validation:Optional
	// +kubebuilder:validation:MaxItems=32
	Users []User `json:"users,omitempty"`

	// NTP defines the NTP configuration for the cluster.
	// +kubebuilder:validation:Optional
	NTP *NTP `json:"ntp,omitempty"`
```

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->