feat: EKS ClusterClass and example (#1296)
**What problem does this PR solve?**:
Adding a ClusterClass and an example file for EKS clusters.

Based on
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1255,
1. with minor changes to fixup after more recent changes
2. change to not rely on
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1252
3. minor refactoring in kustomize to align better with the other
examples

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->

---------

Co-authored-by: Jimmi Dyson <jimmidyson@gmail.com>