test: new LoadBalancer Service e2e test (#1304)
**What problem does this PR solve?**:
Creates a new check in Nutanix e2e tests that reservces an IP in the
Subnet for metallb, creates a Deployment and Service and verifies its
reachable from inside the runner.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->