build: go1.25.1 to fix CVE (#1314)
The majority of changes here are to ensure that the kube-api-linter tool
is built with the same version as golangci-lint, which in turn is built
with the same go language version of the project, otherwise linting will
fail due to language incompatibilities.

**What problem does this PR solve?**:

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->