refactor: Use metallb APIs instead of unstructured (#1317)
Also refactors the apply logic to return early on anything other than a
conflict. Other errors are not necessarily retryable and hence returning
early is better practice.

I decided to do this while I was thinking about introducing the Cilium
load-balancer and looking at using vendored Cilium API types instead of
unstructured - doing this for metallb was nice and simple.

**What problem does this PR solve?**:

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->