refactor: move handler directories (#1321)
**What problem does this PR solve?**:
Moving the handlers and docs around to make it clear when adding new
providers and new handlers if they apply to all providers or only those
that use kubeadm too bootstrap the Nodes.

This better aligns with the changes in
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1297

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->