fix: Replace empty string as Nutanix image name placeholder value
When we update to the next minor release of CAPX (v1.8.x), the empty
string we use as the image name placeholder will be invalid.

This replaces the empty string with a value that will be valid in CAPX
v1.8.x, but is also valid in all previous CAPX versions.