build: handle missing creds Secrets listing CSI images (#1359)
**What problem does this PR solve?**:
The CSI bump
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1320
introduced this change and broke our automation with:
```
PC secret 'nutanix-csi-credentials' in namespace '' not found when usePC is set to true. Please create the secret before installing or upgrading the chart.
```

This works around it by passing dummy values to the automation. 

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->