build: fix list-images for AWS LB controller (#1364)
**What problem does this PR solve?**:
Follow up to
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1341,
I forgot to update this automation after changing the template.

That PR broke list-images, but now with
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1359
there's a preflight that catches these types of errors.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->