fix: webhook panic when management cluster has nil topology (#1372)
**What problem does this PR solve?**:
Fixes a panic in the addon webhook and handles the case when creating a
ClusterClass workload cluster in a non ClusterClass management cluster.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->