fix: set topology webhook matchConditions (#1375)
**What problem does this PR solve?**:
Recently found a bug
https://github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/pull/1372
This adds another level of safety and skips unnecessarily calling this
webhook for clusters not using topology.

**Which issue(s) this PR fixes**:
Fixes #

**How Has This Been Tested?**:
<!--
Please describe the tests that you ran to verify your changes.
Provide output from the tests and any manual steps needed to replicate
the tests.
-->

**Special notes for your reviewer**:
<!--
Use this to provide any additional information to the reviewers.
This may include:
- Best way to review the PR.
- Where the author wants the most review attention on.
- etc.
-->