fix: Copy ClusterClass and Templates only if they do not already exist
It is possible for some resource to exist, but a copy of it cannot be
created. For example, CAPX validation became more strict without
changing its API version. Templates already on the cluster remain
usable. However, they cannot be copied, because the copy does not pass
validation.

Prior to this change, we always copied the template, and ignored the
"already exists" error. Now we first check if the template exist, and
only create it if it does not.