fix: avoid using Helm for legacy agent webhook
The legacy agent release can be installed in any namespace and does not have identifiable labels.
This led to the container being OOMKilled when listing and decoding even a moderate number of Helm releases Secrets.

Instead, this PR checks for a statically named ConfigMap to determine if the legacy agent was installed.